Move create enemy to other class

+++ Game/src/sample/moveAbleObj.java
@@ -1,7 +1,7 @@
 package sample;
 
 public abstract class moveAbleObj extends gameObject{
-    int speed;
+    double speed;
     Direction direction;
 }
